requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.12.0
dnspython==2.8.0
email-validator==2.3.0
execnet==2.1.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-timeout==2.2.0
pytest-xdist==3.5.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.15.0
urllib3==2.5.0
